requirements.txt
asgiref==3.8.1 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
django-cors-headers==4.6.0 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
django-filter==24.3 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
django-multiselectfield==0.1.13 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
django==5.1.3 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
djangorestframework==3.15.2 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
drf-yasg==1.21.8 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
gunicorn==21.2.0 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
inflection==0.5.1 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
mysqlclient==2.2.6 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
numpy==2.1.3 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
packaging==24.2 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
pillow==11.0.0 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
pymysql==1.1.1 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
python-dotenv==1.0.1 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
pytz==2024.2 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
pyyaml==6.0.2 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
sqlparse==0.5.2 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
typing-extensions==4.12.2 ; python_full_version >= "3.10.12" and python_version < "3.11"
tzdata==2024.2 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
wheel==0.45.1 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
whitenoise==6.8.2 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
